Rename and add official patch.xp3
