+++ README.md
@@ -16,11 +16,16 @@ Link: [Fuwanovel Forum](https://forums.fuwanovel.net/topic/24935-hinata-no-tsuki
 ## Repository Outline
 ### Content Files:
 - image :  UI images
-- script :  text and dialogue of story, code commands begin line with `#` and comments begin line with `;`
+- mtl-script : machine translated scripts (to edit and move to script)
+- original-script: japanese scripts before offical update patched
+- script :  translated scripts in progress
+  - text and dialogue of story, code commands begin line with `#` and comments begin line with `;`
 - system :  visual novel engine code files
 - text :  a sort of labelling/reference name for resources
-- update patch :  non-script files from update patch to be released together with translated files in patch.xp3
+- update patch :  non-script files from official update patch to be released together with translated files in patch.xp3
 - utils: code snippets to do what I require. With documentation in their respective folders. 
+
+-official-patch.xp3 : the official patch provided by ko-eda
 
 ## Changes: 
 - UI images translated
