--- utils/patch-update/README.md
@@ -1,12 +1,0 @@
-## Update code/script commands from one file to another without changing text
-
-Script/Code commands in .src start with `;` or `#`.
-
-1. src files with text you want to keep in `old` folder
-2. src files with new code commands in `new` folder
-3. Run `insert_patch.py`
-4. output src file with updated code commands in `out` folder
-
-Purpose:
-https://github.com/Hinata-no-Tsuki-Translations/patch/commit/97baf586ffffdf76ec8d811a79b365b5547e8c5d
-
